Cost function discrepancy
The (fast) new cost function for slr smoothing gives ever so slightly
different values.
Will do proper testing but it is so small that it is possible that it is
numerical. I'm reluctant to write something off as numerical though.
It actually has lot of impact since at convergence, the small
differences cause disagreement on convergence.
This exposed a bug where not the latest approved estimates were returned
at convergence, but rather the latest un-approved, heavily regularised
ditos.